move #35: 폴더 구조와 이름 변경
- 목적
- 초기 개발 시 각 페이지의 독립성을 유지하면서,
- 필요한 경우에만 공통화하는 효율적인 전략
