+++ src/App.jsx
@@ -3,8 +3,8 @@ import './App.scss';
 import Main from './pages/Main';
 import PostId from './pages/PostId';
 import List from './pages/List';
-import PostFromPage from './pages/PostFromPage';
-import PostToPage from './pages/PostToPage';
+import PostIdMessage from './pages/post-id-message-page/PostIdMessage';
+import Post from './pages/post-page/Post';
 
 function App() {
   return (
@@ -12,8 +12,8 @@ function App() {
       <Route path='/' element={<Main />} />
       <Route path='/post/:id' element={<PostId />} />
       <Route path='/list' element={<List />} />
-      <Route path='/post/:id/message' element={<PostFromPage />} />
-      <Route path='/post' element={<PostToPage />} />
+      <Route path='/post/:id/message' element={<PostIdMessage />} />
+      <Route path='/post' element={<Post />} />
     </Routes>
   );
 }
