+++ src/components/post-page/ColorPicker.scss
@@ -1,4 +1,4 @@
-@import '../variables', '../reset';
+@import '../../styles/variables', '../../styles/variables';
 
 @mixin flex-center {
   display: flex;
