+++ src/components/post-page/ImagePicker.scss
@@ -1,4 +1,4 @@
-@import '../variables', '../reset';
+@import '../../styles/variables', '../../styles/reset';
 
 @mixin flex-center {
   display: flex;
