+++ src/pages/post-page/Post.scss
@@ -1,4 +1,4 @@
-@import '../variables', '../reset';
+@import '../../styles/variables', '../../styles/reset';
 
 .background-form {
   width: 320px;
